feat: adopt library-skills convention for agent skill discovery
Move user-facing skills into src/core/trulens/.agents/skills/ so they
ship with the trulens-core package and are auto-discoverable by
`uvx library-skills`. Repo-root skills/ entries become symlinks to
the canonical package location. notebook-execution/ remains at repo
root as contributor-only tooling.

Ref: https://github.com/tiangolo/library-skills
